AETHERION/requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
PyYAML>=6.0
tqdm>=4.64.0
matplotlib>=3.7.0
roboflow>=1.1.0
supervision>=0.14.0
twilio>=8.0.0
firebase-admin>=6.0.0
onnx>=1.14.0
onnxruntime>=1.15.0
